# opt: build partial index predicates as FiltersExprs